+++ doc/changes.md
@@ -3,6 +3,7 @@
 ## next
 
 * FIX: measurement spec uses "country" not "cc" as probe country code filter
+* FIX: measurement spec probe ASN filter was incorrect when tags were not used
 
 ## v0.7.1